Revert "upgrade to python 3.1"
This reverts commit 33a8054eee19d80b5cce94e7f93025809abbf1d1.

# Conflicts:
#	engine/main.h

+++ engine/font.cpp
@@ -48,7 +48,7 @@ namespace Ika {
 
     uint Font::GetGlyphIndex(char c, uint subset) const {
         if (subset >= 0 && subset < _fontFile.NumSubSets()) {
-            return _fontFile.GetSubSet(subset).glyphIndex[(int)c];
+            return _fontFile.GetSubSet(subset).glyphIndex[c];
         } else {
             return 0;
         }
